extraction for batch processing
