+++ deprecated/main/groovy/OrientationConverter.groovy
