+++ deprecated/main/groovy/ReadYaml-complex.groovy
