+++ deprecated/main/groovy/ch/epfl/classes/YamlClasses.groovy
