+++ deprecated/main/groovy/testYamlRW.groovy
