+++ deprecated/main/groovy/test_conditional_gui.groovy
